% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GETINFO.R
\docType{data}
\name{FRED_APIkey}
\alias{FRED_APIkey}
\title{GETINFO from FRED}
\format{
An object of class \code{character} of length 1.
}
\usage{
GETINFO(FRED_APIkey,search_text,realtime_start,realtime_end,limit)
}
\value{

}
\description{
This function allows you retrieve all country's Econ information from FRED(this function requires your own api key from FRED)
}
\details{
GETINFO
}
\examples{
GETINFO(FRED_APIkey="43feed24d714332ac6ad5110919b6556", realtime_start="2018-01-01",realtime_end="2020-09-01",search_text="GDP+Japan",limit="50")
}
\author{
Jiatong Li
}
\keyword{datasets}
