% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gym.R
\name{gym}
\alias{gym}
\title{Get information of gyms from yelp using your own yelp api key}
\usage{
gym(api1="your client id ",api_key1= "your client_secret",i="City,State")
}
\value{

}
\description{
simply typing the city name you want to check and your yelp api key, the function will give you all information (address,rate, etc) of 50 gym in that city.
}
\details{
gym
}
\examples{
gym(api1="q-f2lLVGZkys5AVkVPWsuQ",api_key1 =  "tLcN4rX8BzXZTSE5fYKHntHgSbjHilyGQV8VoXq0zGgiwP7RM8rS0PCOuRyRpn6UkMsEE9gnXWqEs2V32IreXfwVjao7iiAgKeJQ85u2n_mbuZn3-Tuq3R7qs4CPX3Yx",i="San Jose,CA")
}
\author{
Jiatong Li
}
